fix: test errors

+++ src/services/deleteUser.service.ts
@@ -4,18 +4,13 @@ import { User } from "../entities/user.entity";
 const deleteUserService = async (userId: string) => {
   const userRepository = AppDataSource.getRepository(User);
 
-  const user = userRepository.findOneBy({ id: userId });
+  const user = await userRepository.findOneBy({ id: userId });
 
   if (!user) {
     throw new Error("User not found");
   }
 
-  await userRepository
-    .createQueryBuilder()
-    .delete()
-    .from(User)
-    .where("id = :id", { id: userId })
-    .execute();
+  await userRepository.delete({ id: userId });
 };
 
 export default deleteUserService;
